Add notice comment to formatted code blocks
If `prettier-github` formats a code block, this will be noted in a comment
including instructions to disable the bot and the original code block.

Note that this changes the whitespace insertion behaviour from always inserting
a newline after code blocks (default behaviour of `extract-gfm`'s `injectBlocks`
method) to never inserting one.